# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API Documentation",
        "description": "API documentation for Film management",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/filme": {
            "get": {
                "tags": [
                    "Filme"
                ],
                "summary": "Liste aller Filme oder Suche nach Film",
                "operationId": "f85398ce9a7441b1dd51fa569f4fdaed",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Suchbegriff",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste von Filmen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Filme"
                ],
                "summary": "Erstelle einen neuen Film",
                "operationId": "9004a323d5fd1e379e9cbac810c7b74a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilmInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Film erfolgreich erstellt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/filme/{id}": {
            "get": {
                "tags": [
                    "Filme"
                ],
                "summary": "Zeige einen Film ",
                "operationId": "db70560816bd8441594520a8b0d1e1e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Films",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Filme"
                ],
                "summary": "Aktualisiere einen Film",
                "operationId": "0e77d5494d0885e04df4f2020654750f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Films",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilmInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Film aktualisiert",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Filme"
                ],
                "summary": "Lösche einen Film",
                "operationId": "386034c9b596019cdb84dbca066989c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Films",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Film gelöscht"
                    },
                    "404": {
                        "description": "Film nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Film wird noch verwendet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film kann nicht gelöscht werden, da er verwendet wird."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "foreign_key_constraint_violation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Film": {
                "required": [
                    "id",
                    "filmtitel",
                    "bild_url"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "filmtitel": {
                        "type": "string",
                        "example": "Oppenheimer"
                    },
                    "bild_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://www.themoviedb.org/t/p/w1280/8Gxv8gSFCU0XGDykEGv7zR1n2ua.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "FilmInput": {
                "required": [
                    "filmtitel",
                    "bild_url"
                ],
                "properties": {
                    "filmtitel": {
                        "type": "string",
                        "example": "Oppenheimer"
                    },
                    "bild_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://www.themoviedb.org/t/p/w1280/8Gxv8gSFCU0XGDykEGv7zR1n2ua.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}